+++ src/console/controllers/FieldsController.php
@@ -251,6 +251,8 @@ MD, $fields->keys()->join(',')));
             return ExitCode::OK;
         }
 
+        $projectConfig = Craft::$app->getProjectConfig();
+
         $migrationPaths = [];
         $relationFieldHandles = [];
 
@@ -334,6 +336,10 @@ MD, $fields->keys()->join(',')));
                     $this->mergeFields($persistentField, $outgoingField, $usagesByField[$outgoingField->id], $migrationPaths);
                     $this->output();
                 });
+
+            // flush out the project config in case the command is aborted early
+            // (https://github.com/craftcms/cms/issues/16198)
+            $projectConfig->flush();
         }
 
         if (!empty($migrationPaths)) {